requirements.txt
streamlit
pandas
openpyxl
matplotlib
Pillow
fpdf
streamlit-drawable-canvas
xlsxwriter
gspread
google-auth